Refactor to move types in _layout to dedicated types directory
