+++ components/Workout/Form/Exercise/Sets/Sets/WeightThrows/Set.tsx
@@ -15,7 +15,7 @@ import Swipeable from 'react-native-gesture-handler/Swipeable';
 import FontAwesome6 from '@expo/vector-icons/FontAwesome6';
 import Ionicons from '@expo/vector-icons/Ionicons';
 import { errorHaptic, lightHaptic } from "@/utils/haptics/haptics";
-import { FormValues } from "@/app/(tabs)/_layout";
+import { FormValues } from "@/types/commonTypes";
 
 import { styles } from "../styles";
